Foro part 2 con gestion de foro para el adm

+++ BACKEND/aantik/src/main/java/com/aantik/demo/GaboPrivado/CrearPostRepository.java
@@ -10,7 +10,7 @@ import com.aantik.demo.GaboPrivado.*;;
 
 
 @Repository
-public interface CrearPostRepository extends JpaRepository<CrearPosts, Long>{
+public interface CrearPostRepository extends JpaRepository<CrearPosts, Integer>{
 
     
 }
